requirements.txt
fastapi
uvicorn
websockets
anthropic
langchain
langchain-anthropic
langchain-community
chromadb
sentence-transformers
rank-bm25
arxiv
wikipedia
tavily-python
redis
sqlalchemy
streamlit
altair
python-dotenv
structlog
tenacity
pydantic
aiosqlite
pandas
langchain-ollama
langserve
